# docs/swagger/gonuxt.openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "rpc_create_user.proto",
        "version": "version not set"
    },
    "tags": [
        {
            "name": "GoNuxt"
        }
    ],
    "paths": {
        "/v1/create_user": {
            "post": {
                "operationId": "GoNuxt_CreateUser",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pbCreateUserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "An unexpected error response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rpcStatus"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/pbCreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "GoNuxt"
                ]
            }
        },
        "/v1/login_user": {
            "post": {
                "operationId": "GoNuxt_LoginUser",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pbLoginUserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "An unexpected error response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rpcStatus"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/pbLoginUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "GoNuxt"
                ]
            }
        },
        "/v1/refresh_token": {
            "post": {
                "operationId": "GoNuxt_RefreshToken",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pbRefreshTokenResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "An unexpected error response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rpcStatus"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/pbRefreshTokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "GoNuxt"
                ]
            }
        },
        "/v1/update_user": {
            "patch": {
                "operationId": "GoNuxt_UpdateUser",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pbUpdateUserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "An unexpected error response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/rpcStatus"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/pbUpdateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "GoNuxt"
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "pbCreateUserRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "pbCreateUserResponse": {
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/pbUser"
                    }
                }
            },
            "pbLoginUserRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "pbLoginUserResponse": {
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/pbUser"
                    },
                    "session_id": {
                        "type": "string"
                    },
                    "access_token": {
                        "type": "string"
                    },
                    "access_token_expires_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "refresh_token": {
                        "type": "string"
                    },
                    "refresh_token_expires_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "pbRefreshTokenRequest": {
                "type": "object",
                "properties": {
                    "refresh_token": {
                        "type": "string"
                    }
                }
            },
            "pbRefreshTokenResponse": {
                "type": "object",
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "access_token_expires_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "pbUpdateUserRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "pbUpdateUserResponse": {
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/pbUser"
                    }
                }
            },
            "pbUser": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "protobufAny": {
                "type": "object",
                "properties": {
                    "@type": {
                        "type": "string"
                    }
                },
                "additionalProperties": {}
            },
            "rpcStatus": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/protobufAny"
                        }
                    }
                }
            }
        }
    }
}